Make syslog dependency optional
Sometimes it is nice to be able to run email2pdf without requiring that
a syslog socket first be available or to be able to run it even if we
don't know the location of the syslog socket. This commit simply makes
the syslog requirement optional so that email2pdf only outputs to
stderr if it can't find a syslog socket.