Use list for FORMATTED_HEADERS_TO_INCLUDE
Using a frozenset for headers to include might cause the
order of these headers to change with each invocation of
email2pdf. (And it does on my system.) I would therefore
suggest using a data structure that has an order. This
guarantees that iterating over headers will always yield
them in the same order.